% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Genotypes, SNP info, kinship and phenotypes.}
\format{A list with 4 components:
\describe{
  \item{X}{matrix of imputed genotypes}
  \item{Y}{vector of phenotypes}
  \item{K}{kinship matrix}
  \item{snp_info}{SNP coordinates}
}}
\usage{
example_data
}
\description{
A dataset used as example for the mlmm function.
}
\keyword{datasets}
