% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmm_cof.r
\name{mlmm_cof}
\alias{mlmm_cof}
\title{MLMM_COF}
\usage{
mlmm_cof(Y, X, cofs, K, nbchunks, maxsteps, thresh = NULL)
}
\arguments{
\item{Y}{phenotypes, a vector of length m, with names(Y)=individual names}

\item{X}{genotypes, a n by m matrix, where n=number of individuals, m=number of SNPs, with rownames(X)=individual names, and colnames(X)=SNP names}

\item{cofs}{covariates, a n by p matrix, where n=number of individuals, p=number of covariates in the matrix (e.g. PC axes)}

\item{K}{kinship, a n by n matrix, with rownames(K)=colnames(K)=individual names}

\item{nbchunks}{an integer defining the number of chunks of X to run the analysis, allows to decrease the memory usage ==> minimum=2, increase it if you do not have enough memory}

\item{maxsteps}{maximum number of steps desired in the forward approach. The forward approach breaks automatically once the pseudo-heritability is close to 0, however to avoid doing too many steps in case the pseudo-heritability does not reach a value close to 0, this parameter is also used. It's value must be specified as an integer >= 3}

\item{thresh}{threshold}
}
\value{
results
}
\description{
MLMM_COF
}
\author{
V. Segura & B. J. Vilhjalmsson
}
