% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdata-data.R
\docType{data}
\name{pdata}
\alias{pdata}
\title{pdata for examples and testthat()
pdata has 7 variables and var1 is the response variable, var2 is a
continuous explanatory variable,
treatment is another explanatory variable, miss and censor are
indicator for missing and censored,
geneid and sid represents gene id and subject id respectively.}
\format{An object of class \code{data.frame} with 100 rows and 7 columns.}
\usage{
data(pdata)
}
\description{
pdata for examples and testthat()
pdata has 7 variables and var1 is the response variable, var2 is a
continuous explanatory variable,
treatment is another explanatory variable, miss and censor are
indicator for missing and censored,
geneid and sid represents gene id and subject id respectively.
}
\keyword{datasets}
