% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mlmm.r
\name{plot_fwd_GWAS}
\alias{plot_fwd_GWAS}
\title{Plot}
\usage{
plot_fwd_GWAS(x, step, snp_info, pval_filt)
}
\arguments{
\item{x}{x}

\item{step}{step}

\item{snp_info}{snp_info}

\item{pval_filt}{pval_filt}
}
\description{
Plot
}
\author{
V. Segura & B. J. Vilhjalmsson
}

