% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mlmm.r
\name{plot_fwd_region}
\alias{plot_fwd_region}
\title{Plot}
\usage{
plot_fwd_region(x, step, snp_info, pval_filt, chrom, pos1, pos2)
}
\arguments{
\item{x}{x}

\item{step}{step}

\item{snp_info}{snp_info}

\item{pval_filt}{pval_filt}

\item{chrom}{chrom}

\item{pos1}{pos1}

\item{pos2}{pos2}
}
\description{
Plot
}
\author{
V. Segura & B. J. Vilhjalmsson
}

