% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mlmm.r
\name{plot_GWAS}
\alias{plot_GWAS}
\title{Plot}
\usage{
plot_GWAS(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{arguments to be passed to \code{\link[graphics]{plot}}, such as \code{main}}
}
\description{
Plot
}
\author{
V. Segura & B. J. Vilhjalmsson
}
