% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mlmm.r
\name{plot_region}
\alias{plot_region}
\title{Plot}
\usage{
plot_region(x, chrom, pos1, pos2)
}
\arguments{
\item{x}{x}

\item{chrom}{chrom}

\item{pos1}{pos1}

\item{pos2}{pos2}
}
\description{
Plot
}
\author{
V. Segura & B. J. Vilhjalmsson
}
