% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mlmm.r
\name{qqplot_fwd_GWAS}
\alias{qqplot_fwd_GWAS}
\title{Plot}
\usage{
qqplot_fwd_GWAS(x, nsteps)
}
\arguments{
\item{x}{x}

\item{nsteps}{nsteps}
}
\description{
Plot
}
\author{
V. Segura & B. J. Vilhjalmsson
}

