% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mlmm.r
\name{plot_opt_GWAS}
\alias{plot_opt_GWAS}
\title{Plot}
\usage{
plot_opt_GWAS(x, opt, snp_info, pval_filt, ...)
}
\arguments{
\item{x}{x}

\item{opt}{opt}

\item{snp_info}{snp_info}

\item{pval_filt}{pval_filt}

\item{...}{arguments to be passed to \code{\link[graphics]{plot}}, such as \code{main}}
}
\description{
Plot
}
\author{
V. Segura & B. J. Vilhjalmsson
}
